+++ R/run_sim_parallel.R
@@ -156,8 +156,10 @@ run_sim_parallel <- function(arm_list=c("int","noint"),
   )
   
   if(  length(matched_list_forbidden)>0){
-    stop(paste0("Name(s) or object `", list_forbidden_names[matched_list_forbidden],"` defined belong to the list of forbidden names, which can cause issues in the model.
-         Please remove or rename those names. See run_sim or run_sim_parallel for a full list of these names.\n"))
+    stop(paste0("Name(s) or object: `", list_forbidden_names[matched_list_forbidden],"` defined belong to the list of forbidden names, which can cause issues in the model.
+    This is likely caused by an object being declared in the parent/global environment (e.g., `i`) or in the inputs. 
+    You can use functions like find('i') to help you locate them. 
+    Please remove or rename those objects. See run_sim or run_sim_parallel for a full list of these names.\n"))
   }
   
   
@@ -306,7 +308,7 @@ run_sim_parallel <- function(arm_list=c("int","noint"),
           input_list_sens <- c(input_list_sens, list.sensitivity_inputs[[1]])
         } else{
         if ((!is.null(names(list.sensitivity_inputs[[1]]))) & sens==1) {
-          warning("Item ", names(list.sensitivity_inputs), " is named. It is advised to assign unnamed objects if they are going to be processed in the model, as they could generate errors.\n")
+          warning("Item ", names(list.sensitivity_inputs), " is named. It is advised to assign unnamed objects or use lists if they are going to be processed in the model, as they could generate errors.\n")
         }
         input_list_sens <- c(input_list_sens,list.sensitivity_inputs)
         }
@@ -405,7 +407,7 @@ run_sim_parallel <- function(arm_list=c("int","noint"),
       
       # Run engine ----------------------------------------------------------
   
-        final_output <- RDICE:::run_engine(arm_list=arm_list,
+        final_output <- run_engine(arm_list=arm_list,
                                         common_pt_inputs=common_pt_inputs,
                                         unique_pt_inputs=unique_pt_inputs,
                                         input_list = input_list,
@@ -443,7 +445,6 @@ run_sim_parallel <- function(arm_list=c("int","noint"),
     print(paste0("Time to run analysis ", sens,": ",  round(proc.time()[3]- start_time_analysis[3] , 2 ), "s"))
     
   }
-  }, enable=TRUE) 
   print(paste0("Total time to run: ",  round(proc.time()[3]- start_time[3] , 2), "s"))
   
 
@@ -462,6 +463,8 @@ run_sim_parallel <- function(arm_list=c("int","noint"),
   results <- output_sim
   
   RNGkind(rng_kind_store)
+  
+  }, enable=TRUE) 
   
   return(results)